Namespace MigrationSpecStubs to Spec::Support::MigrationStubs
