--- spec/support/migration_spec_stubs.rb
@@ -1,9 +1,0 @@
-module MigrationSpecStubs
-  def self.reserved_stub
-    Class.new(ActiveRecord::Base) do
-      self.table_name = "reserves"
-      self.inheritance_column = :_type_disabled # disable STI
-      serialize :reserved
-    end
-  end
-end
